Record intent on span
Fix issues with middleware order
Document
